+++ src/App.tsx
@@ -1,13 +1,11 @@
+import { TodoPage } from 'pages';
 import React from 'react';
-
-import ToDoPage from './ToDoPage';
-
 import './App.css';
 
 function App() {
   return (
     <main className="App">
-      <ToDoPage />
+        <TodoPage />
     </main>
   );
 }
