+++ src/pages/TodoPage/index.css
@@ -1,0 +1,8 @@
+.ToDo__container {
+  border: 1px solid rgba(0, 0, 0, 0.13);
+  border-radius: 8px;
+  width: 500px;
+  margin-top: 5rem;
+  padding: 24px;
+  box-shadow: 2px 2px 1px rgba(0, 0, 0, 0.09), 3px 2px 3px rgba(0, 0, 0, 0.05);
+}
